R6 Class for images associated with a recipe

Recipe images are not required, but helps the viewer visualize what the recipe is supposed to look like. An Image can either be a local file or a URL that points to the file. In the case of the latter, a method is included to download the image.

Details

This class also includes a method to "plot" the image given it is a recognized image format (.jpeg or .png as of now). This helps view which images to add to a Recipe object without having to open up the image outside of RStudio.

Fields

file

Scalar character vector; the path to an image file. This image needs to be of JPEG or PNG format.

url

Scalar character vector; the URL pointing to an image. Also needs to be of JPEG or PNG format.

name

Scalar character vector; the name of the image, for reference. You can specify directly or omit and the name will be determined using the image file path or URL.

Creating the image object

An Image is an R6 object, and it can be created using Image$new(). A file path or URL pointing to a valid image is required upon initialization; the name of the image can be determined automatically if not provided.

See also