R6 Class providing standard format for recipes

This class will encapsulate various recipe elements into a standard, consistent format. Once the object is built, it will be easier to work with when incorporating in other projects, i.e., a website.

Fields

name

Scalar character vector; the name of the recipe.

summary

Character vector; a brief description of what the recipe makes.

servings

Numeric vector of length 1 or 2; if only one value provided, the number of servings the recipe produces. If 2 values provided, the minimum and maximum number of servings that the recipe produces.

time

(Optionally named) list of lubridate::duration objects (or objects that can be converted to lubridate::duration objects, i.e., the total number of seconds (since lubridate::duration objects are measured in seconds) or string representations of durations, e.g., "14 hours 20 minutes"); you can provide a list with only one element to represent the total recipe time, or you can provide a named list with elements that represent different time components, e.g., "active", "prep", etc.. Providing an unnamed lists with multiple elements will return an error because it is unclear which element corresponds to which time component.

images

A list of Image objects associated with this Recipe.

ingredients

A list of Ingredient objects associated with this Recipe.

equipment

A list of Equipment objects associated with this Recipe.

instructions

A list of Instruction objects associated with this Recipe.

Creating the recipe object

A Recipe is an R6 object, and it can be created using Recipe$new(). A Recipe has several different fields that you could pass as arguments when initializing, however, you only need to provide the name when first initializing; other fields can be added at a later time.

Uses for the recipe object

This class was created mainly to store recipes in a consistent format. Recipe fields can be manually provided, but I've been webscraping and adding fields that way. Once fully built, different recipe fields can be accessed using the $ operator.

See also

Examples

# An example recipe for spaghetti sauce:
r <- Recipe$new('Spaghetti Sauce')
print(r)
#> <Recipe>
#> 
#>   Spaghetti Sauce
#> 
#>   NA
#> 
#>   Servings: 
#>   Total Time:  
#> 
#>   Image Count: 0
#>   Ingredient Count: 0
#>   Equipment Count: 0
#>   Instruction Count: 0

# Add a summary:
r$summary <- paste(
  'This is my grandmother\'s secret recipe for her award-winning spaghetti',
  'sauce. It has been passed down for generations and dates back to the',
  '1700\'s!'
)
print(r)
#> <Recipe>
#> 
#>   Spaghetti Sauce
#> 
#>   This is my grandmother's secret recipe for her award-winning
#>   spaghetti sauce. It has been passed down for generations and dates
#>   back to the 1700's!
#> 
#>   Servings: 
#>   Total Time:  
#> 
#>   Image Count: 0
#>   Ingredient Count: 0
#>   Equipment Count: 0
#>   Instruction Count: 0

# Other elements are additional R6 classes. See the help page of each for
# additional information.